refactor: 유틸 클래스 패키지 분리
