+++ src/main/java/nextstep/ladder/util/ConnectStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.util;
 
 public interface ConnectStrategy {
     public boolean isConnected();
