+++ src/main/java/nextstep/ladder/util/RandomConnectStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.util;
 
 import java.util.Random;
